Make timer get monotonic time per cycle and sidebar improvements
Before, the timer relied on a consistent timeout provided by
GLib.timeout_add. This commit changes the timer to directly get the
monotonic time each time it's updated, improving consistency.

This commit also properly makes the sidebar reappear at the last state
it was in when the timer is stopped.